pyproject.toml
[project]
name = "autodebugger-watsonx"
version = "1.0.0"
description = "AI-powered automatic code debugger using IBM WatsonX and Streamlit"
authors = [
    {name = "Ruslan Magana", email = "contact@ruslanmv.com"}
]
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.9,<3.13"
keywords = ["debugging", "watsonx", "ai", "code-analysis", "streamlit", "ibm"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Debuggers",
    "Topic :: Software Development :: Quality Assurance",
    "Typing :: Typed",
]

dependencies = [
    "streamlit>=1.31.0,<2.0.0",
    "pandas>=2.0.0,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "requests>=2.31.0,<3.0.0",
    "ibm-watson-machine-learning>=1.0.335,<2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0.0",
    "pytest-cov>=4.1.0,<5.0.0",
    "black>=24.0.0,<25.0.0",
    "ruff>=0.2.0,<1.0.0",
    "mypy>=1.8.0,<2.0.0",
    "isort>=5.13.0,<6.0.0",
    "pre-commit>=3.6.0,<4.0.0",
    "types-requests>=2.31.0,<3.0.0",
]

[project.urls]
Homepage = "https://ruslanmv.com"
Documentation = "https://github.com/ruslanmv/Autodebugger-with-multiple-chatbots#readme"
Repository = "https://github.com/ruslanmv/Autodebugger-with-multiple-chatbots"
Issues = "https://github.com/ruslanmv/Autodebugger-with-multiple-chatbots/issues"

[project.scripts]
autodebugger = "autodebugger.app:main"

[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["autodebugger"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | backup
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["backup/*", "*.ipynb"]

[tool.ruff]
line-length = 100
target-version = "py39"
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "__pycache__",
    "backup",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "C",      # flake8-comprehensions
    "B",      # flake8-bugbear
    "UP",     # pyupgrade
    "N",      # pep8-naming
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "TID",    # flake8-tidy-imports
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
    "C901",   # too complex
    "W191",   # indentation contains tabs
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
strict_concatenate = true
disallow_subclassing_any = false
disallow_any_generics = false
exclude = [
    'backup/',
    'build/',
    'dist/',
]

[[tool.mypy.overrides]]
module = [
    "streamlit.*",
    "ibm_watson_machine_learning.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--strict-config",
    "--cov=autodebugger",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["autodebugger"]
omit = [
    "*/tests/*",
    "*/backup/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
